pyproject.toml
[project]
name = "ente_tools"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # API only
    "pynacl>=1.5.0",
    "httpx>=0.28.1",
    "pydantic>=2.10.5",
    "keyring>=25.6.0",
    # CLI (+ pydantic)
    "platformdirs>=4.3.6",
    "rich>=13.9.4",
    "typer>=0.15.1",
    "typer-config[toml]>=1.4.2",
    "humanize>=4.11.0",
]
license = "Apache-2.0"
license-files = ["LICENSE"]
authors = [
  { name = "Mark Scannell", email = "mescanne@gmail.com" }
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]

[project.urls]
Homepage = "https://github.com/pypa/sampleproject"
Issues = "https://github.com/pypa/sampleproject/issues"

[project.scripts]
ente-tool = "ente_tools.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

# [tool.hatch.build.targets.wheel]
# packages = ["src"]

[tool.ruff]
line-length = 120
lint.select = ["ALL"]
lint.ignore = [
    # Allow f-string in logging
    # "G004",
    # No links required on TODO
    "TD003",
    # Ignore TODO
    "FIX002",
    # Ignore commented-out code
    # "ERA001",
    # Ignore trailing whitespace
    # "W291",
]
lint.fixable = ["ALL"]

[tool.pyright]
venvPath = "."
venv = ".venv"
